+++ crates/px-pipeline/src/project/mod.rs
@@ -1,1 +1,3 @@
+pub mod grid_mosiac;
 pub mod register;
+pub mod spiral_mosiac;