+++ crates/px/src/commands/mod.rs
@@ -12,8 +12,6 @@ pub(crate) use observation::preview::preview_observation;
 pub(crate) use profile::info::show_profile;
 pub(crate) use profile::init::init_profile;
 pub(crate) use profile::scan::scan_profile;
-pub(crate) use project::add::add_project_observation;
-pub(crate) use project::align::align_project;
 pub(crate) use project::init::init_project;
 pub(crate) use project::list::list_projects;
 pub(crate) use project::sample::create_project_samples;