+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 * Reworked the meta models in `px-fits` and how headers and capture dates were parsed
 * Change `--bias` to be optional on `px master flat` and allow the bias to be found in the index
 * `px obs calibrate` no longer enforces bias, dark, or flat to support smart scopes that just need debayering
+* Rework project config and implement `single` and `spiral_mosiac` stacking for projects
 
 ## 0.1.3